chore: friday signing progress, pt2
